requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
email-validator
passlib[bcrypt]
pytest
pytest-asyncio
httpx
python-dotenv
